enable buffered output in tests
