+++ src/test/scala/redis/RedisSpec.scala
@@ -7,16 +7,13 @@ import java.util.concurrent.atomic.AtomicInteger
 
 import akka.actor.ActorSystem
 import akka.testkit.TestKit
-import org.apache.logging.log4j.scala.Logger
 import org.scalatest.BeforeAndAfterAll
 
 import scala.collection.JavaConverters._
 import scala.concurrent.Future
 import scala.io.Source
-import scala.reflect.io.File
 import scala.sys.process.{ProcessIO, _}
 import scala.util.Try
-import scala.util.control.NonFatal
 
 object RedisServerHelper {
   val redisHost = "127.0.0.1"
@@ -79,15 +76,6 @@ abstract class RedisHelper extends TestKit(ActorSystem()) with TestBase with Bef
     log.debug(s"redis was started on $port")
   }
 
-}
-
-case class RedisVersion(major: Int, minor: Int, patch: Int) extends Ordered[RedisVersion] {
-
-  import scala.math.Ordered.orderingToOrdered
-
-  override def compare(that: RedisVersion): Int =
-    (this.major, this.minor, this.patch)
-      .compare((that.major, that.minor, that.patch))
 }
 
 abstract class RedisStandaloneServer extends RedisHelper {
@@ -252,60 +240,4 @@ abstract class RedisClusterClients() extends RedisHelper {
     fileStream.iterator().asScala.foreach(Files.delete)
     Files.delete(fileDir.toPath)
   }
-}
-
-import redis.RedisServerHelper._
-
-class RedisProcess(val port: Int) {
-  var server: Process         = null
-  val cmd                     = s"${redisServerCmd} --port $port ${redisServerLogLevel}"
-  protected val log           = Logger(getClass)
-  protected val processLogger = ProcessLogger(line => log.debug(line), line => log.error(line))
-
-  def start(): Unit = {
-    log.debug(s"starting $this")
-    if (server == null)
-      server = Process(cmd).run(processLogger)
-  }
-
-  def stop(): Unit = {
-    log.debug(s"stopping $this")
-    if (server != null) {
-      try {
-        val out = new Socket(redisHost, port).getOutputStream
-        out.write("SHUTDOWN NOSAVE\n".getBytes)
-        out.flush()
-        out.close()
-      } catch {
-        case NonFatal(e) => log.error(s"couldn't stop $this", e)
-      } finally {
-        server.destroy()
-        server = null
-      }
-    }
-  }
-
-  override def toString: String = s"RedisProcess($port)"
-}
-
-class SentinelProcess(masterName: String, masterPort: Int, port: Int) extends RedisProcess(port) {
-  val sentinelConfPath = {
-    val sentinelConf =
-      s"""
-         |sentinel monitor $masterName $redisHost $masterPort 2
-         |sentinel down-after-milliseconds $masterName 5000
-         |sentinel parallel-syncs $masterName 1
-         |sentinel failover-timeout $masterName 10000
-            """.stripMargin
-
-    val sentinelConfFile = File.makeTemp("rediscala-sentinel", ".conf")
-    sentinelConfFile.writeAll(sentinelConf)
-    sentinelConfFile.path
-  }
-
-  override val cmd = s"${redisServerCmd} $sentinelConfPath --port $port --sentinel $redisServerLogLevel"
-}
-
-class SlaveProcess(masterPort: Int, port: Int) extends RedisProcess(port) {
-  override val cmd = s"$redisServerCmd --port $port --slaveof $redisHost $masterPort $redisServerLogLevel"
 }
